services/llm-service/requirements.txt
fastapi
uvicorn[standard]
aiokafka
google-cloud-aiplatform
pydantic>=2.0
redis[asyncio]
python-dotenv
websockets
